Decouple hashing process from database (#25)
# Description

Utilizing a secure cryptographic function that takes in the block slot
right before sending the memo transaction and adding that to the file
buffer for hashsing. This way we can retrieve the block slot from the
confirmed transaction and recreate the hash. The is secure because
Solana has a max of 150 blocks before it is considered expired. This
allows it to be fully verified by having the file to verify and the TX
signature.

- Added skeleton rows for the `/docs/list` data table while the data is
being fetch
- Improved some layout formatting
- Improved error messages for `/api/docs/new` and `/api/docs/sign`
- Added the new cryptographic functions to hash and verify
- Added disabling for the `FileUpload` component
- Added new Solana helper functions to confirm the transaction prior to
storing the information in the database to ensure it was confirmed on
the blockchain
- Getter function for the latest block slot prior to the memo
transaction sending
- Getting the transaction data from the confirmed memo transaction for
the document